Rewrite `<ShortNumber />` as FC and TS
